gcc 4.6 fixes, c++0x fixes

+++ src/AC3DPlugins/obj_editor.cpp
@@ -428,7 +428,7 @@ void	OBJ_editor_init(void)
 	blend_level_var = new TCL_linked_vardv(ac_get_tcl_interp(), "xplane_blend_level", MAX_MULTI_COUNT, xplane_blend_level_cb, NULL, 0.0);
 	anim_type_var = new TCL_linked_varsv(ac_get_tcl_interp(),"xplane_anim_type",MAX_MULTI_COUNT, xplane_anim_type_cb,NULL,"no animation");
 
-	for (int n = 0; n < NUM_KEYFRAMES; ++n)
+	for (intptr_t n = 0; n < NUM_KEYFRAMES; ++n)
 	{
 		char	buf[25];
 		sprintf(buf,"xplane_anim_value%d", n);
